p5-advanced/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
structlog==24.4.0

# LightRAG
lightrag-hku[offline-llm]
networkx>=3.2

# Prometheus metrics
prometheus-fastapi-instrumentator==7.0.2

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
